// src/models/openresponsescreatedevent.ts
/*
 * Code generated by Speakeasy (https://speakeasy.com). DO NOT EDIT.
 * @generated-id: e65287120d25
 */

import * as z from "zod/v4";
import { remap as remap$ } from "../lib/primitives.js";
import { safeParse } from "../lib/schemas.js";
import { Result as SafeParseResult } from "../types/fp.js";
import { SDKValidationError } from "./errors/sdkvalidationerror.js";
import {
  OpenResponsesResult,
  OpenResponsesResult$inboundSchema,
} from "./openresponsesresult.js";

/**
 * Event emitted when a response is created
 */
export type OpenResponsesCreatedEvent = {
  type: "response.created";
  /**
   * Complete non-streaming response from the Responses API
   */
  response: OpenResponsesResult;
  sequenceNumber: number;
};

/** @internal */
export const OpenResponsesCreatedEvent$inboundSchema: z.ZodType<
  OpenResponsesCreatedEvent,
  unknown
> = z.object({
  type: z.literal("response.created"),
  response: OpenResponsesResult$inboundSchema,
  sequence_number: z.int(),
}).transform((v) => {
  return remap$(v, {
    "sequence_number": "sequenceNumber",
  });
});

export function openResponsesCreatedEventFromJSON(
  jsonString: string,
): SafeParseResult<OpenResponsesCreatedEvent, SDKValidationError> {
  return safeParse(
    jsonString,
    (x) => OpenResponsesCreatedEvent$inboundSchema.parse(JSON.parse(x)),
    `Failed to parse 'OpenResponsesCreatedEvent' from JSON`,
  );
}